Moved join logs, forgot to earlier
